refactor(api): [#143] move API resources mod
