+++ src/apis/mod.rs
@@ -1,3 +1,4 @@
 pub mod middlewares;
+pub mod resources;
 pub mod routes;
 pub mod server;
