+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod api;
 pub mod apis;
 pub mod config;
 pub mod databases;
